Feature/mlx rework

+++ libft/ft_strnstr.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-char	*ft_strnstr(const char *haystack, const char *needle, unsigned long len)
+char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
 	unsigned long	i;
 	unsigned long	ii;
